Fix type bug

+++ torch/_dynamo/config.py
@@ -384,7 +384,7 @@ trace_numpy = True
 
 # Default NumPy dtypes when tracing with torch.compile
 # We default to 64bits. For efficiency, one may want to change these to float32
-numpy_default_float = "float64"
+numpy_default_float = "float32"
 numpy_default_complex = "complex128"
 numpy_default_int = "int64"
